moved javascript to an external code file
